requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6

# Base de datos PostgreSQL
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Utilidades
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.31.0
pytest-postgresql==5.0.0

# Testing UI y Performance
selenium==4.15.2
webdriver-manager==4.0.1
locust==2.17.0

# Cobertura y calidad
coverage==7.3.2
pytest-cov==4.1.0
black==23.10.1
flake8==6.1.0